+++ eng/pipelines/ci-template.yml
@@ -36,10 +36,6 @@ steps:
       displayName: Install TypeSpec
       condition: and(succeeded(), ${{ parameters.installTypeSpec }})
 
-    - script: npm install -g @azure-tools/cadl-ranch
-      displayName: Install Cadl Ranch
-      condition: and(succeeded(), ${{ parameters.installCadlRanch }})
-
     - script: npx @azure-tools/typespec-bump-deps package.json packages/typespec-python/package.json
       displayName: Update typespec packages to latest dev version
       workingDirectory: $(Build.SourcesDirectory)/autorest.python/
