+++ eng/pipelines/internal-ci.yml
@@ -30,11 +30,11 @@ steps:
     - task: AzureCLI@2
       displayName: Upload scenario manifest for unbranded
       inputs:
-        azureSubscription: "Cadl Ranch Storage"
+        azureSubscription: "Spector Storage"
         scriptType: "bash"
         scriptLocation: "inlineScript"
-        inlineScript: cadl-ranch upload-coverage --generatorName python --storageAccountName azuresdkcadlranch --generatorVersion $(node -p -e "require('$(Build.SourcesDirectory)/autorest.python/packages/typespec-python/package.json').version") --generatorMode standard
-        workingDirectory: $(Build.SourcesDirectory)/autorest.python/packages/typespec-python/node_modules/@azure-tools/cadl-ranch-specs
+        inlineScript: tsp-spector upload-coverage --generatorName "@typespec/http-client-python" --storageAccountName typespec --generatorVersion $(node -p -e "require('$(Build.SourcesDirectory)/autorest.python/packages/typespec-python/node_modules/@typspec/http-client-python/package.json').version") --containerName coverages --generatorMode standard
+        workingDirectory: $(Build.SourcesDirectory)/autorest.python/packages/typespec-python/node_modules/@azure-tools/azure-http-specs
 
     - script: |
             tox run -e ci
@@ -44,8 +44,8 @@ steps:
     - task: AzureCLI@2
       displayName: Upload scenario manifest for azure
       inputs:
-        azureSubscription: "Cadl Ranch Storage"
+        azureSubscription: "Spector Storage"
         scriptType: "bash"
         scriptLocation: "inlineScript"
-        inlineScript: cadl-ranch upload-coverage --generatorName python --storageAccountName azuresdkcadlranch --generatorVersion $(node -p -e "require('$(Build.SourcesDirectory)/autorest.python/packages/typespec-python/package.json').version") --generatorMode azure
-        workingDirectory: $(Build.SourcesDirectory)/autorest.python/packages/typespec-python/node_modules/@azure-tools/cadl-ranch-specs
+        inlineScript: tsp-spector upload-coverage --generatorName "@azure-tools/typespec-python" --storageAccountName typespec --generatorVersion $(node -p -e "require('$(Build.SourcesDirectory)/autorest.python/packages/typespec-python/package.json').version") --containerName coverages --generatorMode azure
+        workingDirectory: $(Build.SourcesDirectory)/autorest.python/packages/typespec-python/node_modules/@azure-tools/azure-http-specs
