+++ eng/pipelines/publish-release.yml
@@ -31,9 +31,6 @@ extends:
                 - script: npm install -g pnpm@9.5.0
                   displayName: Install PNPM 9.5.0
 
-                - script: npm install -g @azure-tools/cadl-ranch
-                  displayName: Install cadl ranch
-
                 - script: pnpm install
                   displayName: Install dependencies
