+++ packages/typespec-python/package.json
@@ -74,9 +74,8 @@
     "@azure-tools/typespec-azure-rulesets": "~0.49.0",
     "@azure-tools/typespec-autorest": "~0.49.0",
     "@azure-tools/typespec-client-generator-core": "~0.49.0",
-    "@azure-tools/cadl-ranch-expect": "~0.15.7",
-    "@azure-tools/cadl-ranch-specs": "~0.39.6",
-    "@azure-tools/cadl-ranch": "~0.16.2",
+    "@azure-tools/azure-http-specs": "0.1.0-alpha.4",
+    "@typespec/http-specs": "0.1.0-alpha.5",
     "@types/js-yaml": "~4.0.5",
     "@types/node": "~22.5.4",
     "@types/yargs": "~17.0.33",
