+++ packages/typespec-python/scripts/eng/regenerate.ts
@@ -12,7 +12,8 @@ const exec = promisify(execCallback);
 
 // Get the directory of the current file
 const PLUGIN_DIR = resolve(fileURLToPath(import.meta.url), "../../../");
-const CADL_RANCH_DIR = resolve(PLUGIN_DIR, "node_modules/@azure-tools/cadl-ranch-specs/http");
+const AZURE_HTTP_SPECS = resolve(PLUGIN_DIR, "node_modules/@azure-tools/azure-http-specs/specs");
+const HTTP_SPECS = resolve(PLUGIN_DIR, "node_modules/@typespec/http-specs/specs");
 interface TspCommand {
     outputDir: string;
     command: string;
@@ -111,7 +112,8 @@ function toPosix(dir: string): string {
 }
 
 function getEmitterOption(spec: string): Record<string, string>[] {
-    const relativeSpec = toPosix(relative(CADL_RANCH_DIR, spec));
+    const specDir = spec.includes("azure") ? AZURE_HTTP_SPECS : HTTP_SPECS;
+    const relativeSpec = toPosix(relative(specDir, spec));
     const key = relativeSpec.includes("resiliency/srv-driven/old.tsp") ? relativeSpec : dirname(relativeSpec);
     const result = EMITTER_OPTIONS[key] || [{}];
     return Array.isArray(result) ? result : [result];
@@ -205,7 +207,8 @@ async function getSubdirectories(baseDir: string, flags: RegenerateFlags): Promi
 }
 
 function defaultPackageName(spec: string): string {
-    return toPosix(relative(CADL_RANCH_DIR, dirname(spec)))
+    const specDir = spec.includes("azure") ? AZURE_HTTP_SPECS : HTTP_SPECS;
+    return toPosix(relative(specDir, dirname(spec)))
         .replace(/\//g, "-")
         .toLowerCase();
 }
@@ -259,8 +262,9 @@ async function regenerate(flags: RegenerateFlagsInput): Promise<void> {
         await regenerate({ ...flags, flavor: "unbranded" });
     } else {
         const flagsResolved = { debug: false, flavor: flags.flavor, ...flags };
-        const CADL_RANCH_DIR = resolve(PLUGIN_DIR, "node_modules/@azure-tools/cadl-ranch-specs/http");
-        const subdirectories = await getSubdirectories(CADL_RANCH_DIR, flagsResolved);
+        const subdirectories1 = await getSubdirectories(AZURE_HTTP_SPECS, flagsResolved);
+        const subdirectories2 = await getSubdirectories(HTTP_SPECS, flagsResolved);
+        const subdirectories = [...subdirectories1, ...subdirectories2];
         const cmdList: TspCommand[] = subdirectories.flatMap((subdirectory) =>
             _getCmdList(subdirectory, flagsResolved),
         );
