+++ packages/typespec-python/test/azure/mock_api_tests/conftest.py
@@ -13,9 +13,10 @@ from pathlib import Path
 
 
 def start_server_process():
-    path = Path(os.path.dirname(__file__)) / Path("../../../node_modules/@azure-tools/cadl-ranch-specs")
-    os.chdir(path.resolve())
-    cmd = "cadl-ranch serve ./http"
+    azure_http_path = Path(os.path.dirname(__file__)) / Path("../../../node_modules/@azure-tools/azure-http-specs")
+    http_path = Path(os.path.dirname(__file__)) / Path("../../../node_modules/@typespec/http-specs")
+    os.chdir(azure_http_path.resolve())
+    cmd = f"tsp-spector serve ./specs  {(http_path / 'specs').resolve()}"
     if os.name == "nt":
         return subprocess.Popen(cmd, shell=True)
     return subprocess.Popen(cmd, shell=True, preexec_fn=os.setsid)
