+++ packages/typespec-python/test/azure/requirements.txt
@@ -19,6 +19,7 @@ azure-mgmt-core==1.3.2
 -e ./generated/azure-example-basic
 -e ./generated/azure-resource-manager-common-properties
 -e ./generated/azure-resource-manager-resources
+-e ./generated/azure-payload-pageable
 
 # common test case
 -e ./generated/authentication-api-key
